pyproject.toml
# ============================================================
# ⚒️ TianGong — 天工 AI Agent 修炼平台 项目配置
# ============================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tiangong-mcp"
version = "0.1.0"
description = "⚒️ TianGong — The Celestial Forge for AI Agents. 开源的 AI Agent 修炼平台。"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "JinNing6", email = "tiangong@celestialforge.dev" },
]
keywords = [
    "mcp", "tiangong", "ai-agent", "agent-forge",
    "model-context-protocol", "cultivation", "celestial-forge",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.3.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/JinNing6/TianGong"
Documentation = "https://github.com/JinNing6/TianGong#readme"
Repository = "https://github.com/JinNing6/TianGong"
Issues = "https://github.com/JinNing6/TianGong/issues"

[project.scripts]
tiangong-mcp = "tiangong.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["tiangong"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (由 formatter 处理)
    "B008",  # do not perform function calls in argument defaults (FastMCP 需要)
]

[tool.ruff.lint.isort]
known-first-party = ["tiangong"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
